chore: refactoring rfccard for rfc link popovers
