+++ client/components/A.vue
@@ -1,34 +1,25 @@
 <template>
-  <component
-    :is="
-      isInternal ?
-        isRfcLink ? RFCRouterLink
-        : RouterLink
-      : 'a'
-    "
-    v-bind="sanitisedAnchorProps"
-  >
+  <component :is="isInternal ? RouterLink : 'a'" v-bind="sanitisedAnchorProps">
     <slot />
   </component>
 </template>
 
 <script setup lang="ts">
+/**
+ * An anchor hyperlink that detects whether to use SPA Vue Router or a conventional hyperlink.
+ *
+ * If you want to ALSO detect RFC links and display a preview use AMaybeRFCLink.vue
+ */
 import { computed } from 'vue'
 import { RouterLink } from 'vue-router'
-import RFCRouterLink from './RFCRouterLink.vue'
 import { EXTERNAL_LINK_REL, TARGET_NEW_WINDOW } from '~/utilities/html'
 import type { AnchorProps } from '~/utilities/html'
-import {
-  isInternalLink,
-  isMailToLink,
-  parseMaybeRfcLink
-} from '~/utilities/url'
+import { isInternalLink, isMailToLink } from '~/utilities/url'
 
 const props = defineProps<AnchorProps>()
 
 const isInternal = computed(() => isInternalLink(props.href))
 const isMailTo = computed(() => isMailToLink(props.href))
-const isRfcLink = computed(() => !!parseMaybeRfcLink(props.href))
 
 const sanitisedAnchorProps = computed(() => ({
   ...props,
